+++ Robust.Shared/GameObjects/Systems/MapSystem.cs
@@ -14,11 +14,33 @@ namespace Robust.Shared.GameObjects
             base.Initialize();
 
             SubscribeLocalEvent<MapGridComponent, ComponentRemove>(RemoveHandler);
+            SubscribeLocalEvent<MapGridComponent, ComponentInit>(HandleGridInitialize);
         }
 
         private void RemoveHandler(EntityUid uid, MapGridComponent component, ComponentRemove args)
         {
             _mapManager.OnComponentRemoved(component);
+        }
+
+        private void HandleGridInitialize(EntityUid uid, MapGridComponent component, ComponentInit args)
+        {
+            var msg = new GridInitializeEvent(uid, component.GridIndex);
+            EntityManager.EventBus.RaiseLocalEvent(uid, msg);
+        }
+    }
+
+    /// <summary>
+    /// Raised whenever a grid is being initialized.
+    /// </summary>
+    public sealed class GridInitializeEvent : EntityEventArgs
+    {
+        public EntityUid EntityUid { get; }
+        public GridId GridId { get; }
+
+        public GridInitializeEvent(EntityUid uid, GridId gridId)
+        {
+            EntityUid = uid;
+            GridId = gridId;
         }
     }
 }
